Started adding config file loading.
Started with the reelstrip class, not done yet.

Added pugiXML and libtech libraries as submodules. It is intended
to switch the UnrealEngine project to load a built version of the
support library. So this project will end up as a shared lib and
be loaded from the SuperSlots project. This will avoid having
to copy and paste the code from one project to another. Also since
this project is now using dependencies of its own it is not viable
to copy them all into the Unreal project.